feat(ESSNTL-4056): move functionality inside the filter component
